Make all fields of structs private, add accessors.
As iouring evolves, the definition of structs might change. This has
already happened, when the offset field of io_uring_sqe became a union
of an offset and a second address, to support more IO operations.

To be forward compatible with future changes, we make all of the fields
of structs private and users can only access them through methods.