pad structures to match liburing
25bbcbef3e0a8bfba8044be55d08d5116c51dccd added padding to the sq
and cq structures.

The iou unit tests all fail once liburing is updated to their main
branch, which I tracked down to the commit above.

This patch also bumps the liburing submodule, to keep things in sync.
Although you have in the past expressed a justified preference to
use official releases, a release of liburing is really close and I
don't see another way to contribute the patch to add the padding
without breaking anything if the submodule doesn't move.